refactor(nge-doc): split the providers into domain files
nge-doc.providers.ts grew to ~290 lines across six concerns. It becomes
core/providers/ with one file per domain: provide (engine defaults,
theme, markdown renderer, dark mode), header (brand, navbar), seo (seo
config, edit link), search (build-time index) and labels (the wording
dictionaries, half of the old file). A barrel keeps the public API and
import paths unchanged.
