+++ projects/nge/doc/index.ts
@@ -2,7 +2,7 @@ export * from './src/core/nge-doc'
 export * from './src/core/assets'
 export * from './src/core/manifest'
 export * from './src/core/search'
-export * from './src/core/nge-doc.providers'
+export * from './src/core/providers'
 export * from './src/core/nge-doc.service'
 export * from './src/core/nge-doc-theme.service'
 export * from './src/shared/pages'
