+++ projects/nge/doc/src/core/nge-doc.service.spec.ts
@@ -5,7 +5,7 @@ import { ActivatedRoute, Router } from '@angular/router'
 import { Subject, of } from 'rxjs'
 import { NgeDocLink, NgeDocSettings, NgeDocState } from './nge-doc'
 import { docsFromManifest } from './manifest'
-import { NGE_DOC_NAVBAR, NGE_DOC_SEO } from './nge-doc.providers'
+import { NGE_DOC_NAVBAR, NGE_DOC_SEO } from './providers'
 import { NgeDocService } from './nge-doc.service'
 import { NgeDocSeoService } from './seo.service'
 import { NgeDocSitesLoader } from './sites-loader'
